# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Zinc Search engine API documents https://zincsearch-docs.zinc.dev",
        "title": "Zinc Search engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Zinc Search",
            "url": "https://www.zincsearch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/_aliases": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Add or remove index alias for compatible ES",
                "operationId": "AddOrRemoveESAlias",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/_bulk": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Document"
                ],
                "summary": "ES bulk documents",
                "operationId": "ESBulk",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/_index_template": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "List index teplates",
                "operationId": "ListTemplates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/meta.Template"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Create update index template",
                "operationId": "CreateTemplate",
                "parameters": [
                    {
                        "description": "Template data",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.IndexTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/_index_template/{name}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get index template",
                "operationId": "GetTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.IndexTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Create update index template",
                "operationId": "UpdateTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Template data",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.IndexTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Delete template",
                "operationId": "DeleteTemplate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/_msearch": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search V2 MultipleSearch for compatible ES",
                "operationId": "MSearch",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/{index}/_delete_by_query": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Searches the index and deletes all matched documents",
                "operationId": "DeleteByQuery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.ZincQueryForSDK"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseDeleteByQuery"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
        "/{index}/_search": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search V2 DSL for compatible ES",
                "operationId": "Search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.ZincQueryForSDK"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/{target}/_alias/{target_alias}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Index"
                ],
                "summary": "Get index alias for compatible ES",
                "operationId": "GetESAliases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Index",
                        "name": "target",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Target Alias",
                        "name": "target_alias",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "meta.AggregationAutoDateHistogram": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "format": {
                    "description": "format key_as_string",
                    "type": "string"
                },
                "keyed": {
                    "type": "boolean"
                },
                "minimum_interval": {
                    "description": "minute,hour,day,week,month,quarter,year",
                    "type": "string"
                },
                "time_zone": {
                    "description": "time_zone",
                    "type": "string"
                }
            }
        },
...
        "meta.AggregationResponse": {
            "type": "object",
            "properties": {
                "buckets": {
                    "description": "slice or map"
                },
                "interval": {
                    "description": "support for auto_date_histogram_aggregation",
                    "type": "string"
                },
                "value": {}
            }
        },
        "meta.Aggregations": {
            "type": "object",
            "properties": {
                "aggs": {
                    "description": "nested aggregations",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Aggregations"
                    }
                },
                "auto_date_histogram": {
                    "$ref": "#/definitions/meta.AggregationAutoDateHistogram"
                },
                "avg": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "cardinality": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "count": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "date_histogram": {
                    "$ref": "#/definitions/meta.AggregationDateHistogram"
                },
                "date_range": {
                    "$ref": "#/definitions/meta.AggregationDateRange"
                },
                "histogram": {
                    "$ref": "#/definitions/meta.AggregationHistogram"
                },
                "ip_range": {
                    "description": "TODO: not implemented",
                    "allOf": [
                        {
                            "$ref": "#/definitions/meta.AggregationIPRange"
                        }
                    ]
                },
                "max": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "min": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "range": {
                    "$ref": "#/definitions/meta.AggregationRange"
                },
                "sum": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "terms": {
                    "$ref": "#/definitions/meta.AggregationsTerms"
                },
                "weighted_avg": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                }
            }
        },
...
        "meta.HTTPResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "meta.HTTPResponseDeleteByQuery": {
            "type": "object",
            "properties": {
                "batches": {
                    "type": "integer"
                },
                "deleted": {
                    "type": "integer"
                },
                "failures": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "noops": {
                    "type": "integer"
                },
                "requests_per_second": {
                    "type": "integer"
                },
                "retries": {
                    "$ref": "#/definitions/meta.HttpRetriesResponse"
                },
                "throttled_millis": {
                    "type": "integer"
                },
                "throttled_until_millis": {
                    "type": "integer"
                },
                "time_out": {
                    "type": "boolean"
                },
                "took": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "version_conflicts": {
                    "type": "integer"
                }
            }
        },
...
        "meta.HTTPResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponseTemplate": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                }
            }
        },
...
        "meta.Highlight": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Highlight"
                    }
                },
                "fragment_size": {
                    "type": "integer"
                },
                "number_of_fragments": {
                    "type": "integer"
                },
                "post_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pre_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "meta.Hit": {
            "type": "object",
            "properties": {
                "@timestamp": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                },
                "_index": {
                    "type": "string"
                },
                "_score": {
                    "type": "number"
                },
                "_source": {},
                "_type": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": true
                },
                "highlight": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "meta.Hits": {
            "type": "object",
            "properties": {
                "hits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.Hit"
                    }
                },
                "max_score": {
                    "type": "number"
                },
                "total": {
                    "$ref": "#/definitions/meta.Total"
                }
            }
        },
        "meta.HttpRetriesResponse": {
            "type": "object",
            "properties": {
                "bulk": {
                    "type": "integer"
                },
                "search": {
                    "type": "integer"
                }
            }
        },
...
        "meta.IndexSettings": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/meta.IndexAnalysis"
                },
                "number_of_replicas": {
                    "type": "integer"
                },
                "number_of_shards": {
                    "type": "integer"
                }
            }
        },
...
        "meta.IndexTemplate": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "index_patterns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "priority": {
                    "description": "highest priority is chosen",
                    "type": "integer"
                },
                "template": {
                    "$ref": "#/definitions/meta.TemplateTemplate"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "meta.Mappings": {
            "type": "object",
            "properties": {
                "properties": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Property"
                    }
                }
            }
        },
...
        "meta.SearchResponse": {
            "type": "object",
            "properties": {
                "_shards": {
                    "$ref": "#/definitions/meta.Shards"
                },
                "aggregations": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.AggregationResponse"
                    }
                },
                "error": {
                    "type": "string"
                },
                "hits": {
                    "$ref": "#/definitions/meta.Hits"
                },
                "timed_out": {
                    "type": "boolean"
                },
                "took": {
                    "description": "Time it took to generate the response",
                    "type": "integer"
                }
            }
        },
        "meta.Shards": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                },
                "successful": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "meta.Template": {
            "type": "object",
            "properties": {
                "index_template": {
                    "$ref": "#/definitions/meta.IndexTemplate"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "meta.TemplateTemplate": {
            "type": "object",
            "properties": {
                "mappings": {
                    "$ref": "#/definitions/meta.Mappings"
                },
                "settings": {
                    "$ref": "#/definitions/meta.IndexSettings"
                }
            }
        },
...
        "meta.Total": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Count of documents returned",
                    "type": "integer"
                }
            }
        },
...
        "meta.ZincQueryForSDK": {
            "type": "object",
            "properties": {
                "_source": {
                    "description": "true, false, [\"field1\", \"field2.*\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aggs": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Aggregations"
                    }
                },
                "explain": {
                    "type": "boolean"
                },
                "fields": {
                    "description": "[\"field1\", \"field2.*\", {\"field\": \"fieldName\", \"format\": \"epoch_millis\"}]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from": {
                    "type": "integer"
                },
                "highlight": {
                    "$ref": "#/definitions/meta.Highlight"
                },
                "query": {
                    "$ref": "#/definitions/meta.QueryForSDK"
                },
                "size": {
                    "type": "integer"
                },
                "sort": {
                    "description": "\"_score\", [\"+Year\",\"-Year\", {\"Year\": \"desc\"}, \"Date\": {\"order\": \"asc\"\", \"format\": \"yyyy-MM-dd\"}}\"}]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "type": "integer"
                },
                "track_total_hits": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}