# pkg/routes/routes.go
/* Copyright 2022 Zinc Labs Inc. and Contributors
*
* Licensed under the Apache License, Version 2.0 (the "License");
* you may not use this file except in compliance with the License.
* You may obtain a copy of the License at
*
*     http://www.apache.org/licenses/LICENSE-2.0
*
* Unless required by applicable law or agreed to in writing, software
* distributed under the License is distributed on an "AS IS" BASIS,
* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
* See the License for the specific language governing permissions and
* limitations under the License.
 */

package routes

import (
	"net/http"
	"time"

	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	_ "github.com/zincsearch/zincsearch/docs" // docs is generated by Swag CLI

	"github.com/zincsearch/zincsearch/pkg/config"
	"github.com/zincsearch/zincsearch/pkg/handlers/document"
	"github.com/zincsearch/zincsearch/pkg/handlers/index"
	"github.com/zincsearch/zincsearch/pkg/handlers/search"
	"github.com/zincsearch/zincsearch/pkg/meta"
	"github.com/zincsearch/zincsearch/pkg/meta/elastic"
	"github.com/zincsearch/zincsearch/pkg/zutils"
)

// SetRoutes sets up all gin HTTP API endpoints that can be called by front end
func SetRoutes(r *gin.Engine) {
	r.Use(cors.New(cors.Config{
		AllowOrigins:     []string{"*"},
		AllowMethods:     []string{"GET", "POST", "DELETE", "PUT", "HEAD", "OPTIONS"},
		AllowHeaders:     []string{"Origin", "authorization", "content-type"},
		ExposeHeaders:    []string{"Content-Length"},
		AllowCredentials: true,
		MaxAge:           12 * time.Hour,
	}))

	r.GET("/version", meta.GetVersion)
	r.GET("/healthz", meta.GetHealthz)

	// use ginSwagger middleware to serve the API docs
	if config.Global.ZincSwaggerEnable {
		r.GET("/swagger", func(c *gin.Context) {
			c.Redirect(http.StatusMovedPermanently, "/swagger/index.html")
		})
		r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}

	/**
	 * elastic compatible APIs
	 */

	r.GET("/", ESMiddleware, func(c *gin.Context) {
		zutils.GinRenderJSON(c, http.StatusOK, elastic.NewESInfo(c))
	})
	r.HEAD("/", ESMiddleware, func(c *gin.Context) {
		zutils.GinRenderJSON(c, http.StatusOK, elastic.NewESInfo(c))
	})
	r.GET("/_license", ESMiddleware, func(c *gin.Context) {
		zutils.GinRenderJSON(c, http.StatusOK, elastic.NewESLicense(c))
	})
	r.GET("/_xpack", ESMiddleware, func(c *gin.Context) {
		zutils.GinRenderJSON(c, http.StatusOK, elastic.NewESXPack(c))
	})

	r.POST("/_search", ESMiddleware, IndexAliasMiddleware, search.SearchDSL)
	r.POST("/_msearch", ESMiddleware, IndexAliasMiddleware, search.MultipleSearch)
	r.GET("/:target/_search", ESMiddleware, IndexAliasMiddleware, search.SearchDSL)
	r.POST("/:target/_search", ESMiddleware, IndexAliasMiddleware, search.SearchDSL)
	r.POST("/:target/_msearch", ESMiddleware, IndexAliasMiddleware, search.MultipleSearch)
	r.POST("/:target/_delete_by_query", IndexAliasMiddleware, search.DeleteByQuery)

	r.GET("/:target/_count", ESMiddleware, IndexAliasMiddleware, search.CountDSL)

	r.GET("/_index_template", ESMiddleware, index.ListTemplate)
	r.POST("/_index_template", ESMiddleware, index.CreateTemplate)
	r.PUT("/_index_template/:target", ESMiddleware, index.CreateTemplate)
	r.GET("/_index_template/:target", ESMiddleware, index.GetTemplate)
	r.HEAD("/_index_template/:target", ESMiddleware, index.GetTemplate)
	r.DELETE("/_index_template/:target", ESMiddleware, index.DeleteTemplate)
	// ES Compatible data stream
	r.PUT("/_data_stream/:target", ESMiddleware, elastic.PutDataStream)
	r.GET("/_data_stream/:target", ESMiddleware, elastic.GetDataStream)
	r.HEAD("/_data_stream/:target", ESMiddleware, elastic.GetDataStream)

	r.PUT("/:target", ESMiddleware, index.CreateES)
	r.HEAD("/:target", ESMiddleware, index.Exists)
	r.DELETE("/:target", ESMiddleware, index.Delete)

	r.GET("/_cat/indices", ESMiddleware, index.CatIndices)
	r.GET("/_cat/indices/:target", ESMiddleware, index.CatIndices)
	r.GET("/_cat/aliases", ESMiddleware, index.CatAliases)
	r.GET("/_cat/aliases/:target", ESMiddleware, index.CatAliases)

	r.GET("/:target/_mapping", ESMiddleware, index.GetESMapping)
	r.PUT("/:target/_mapping", ESMiddleware, index.SetMapping)

	r.GET("/:target/_settings", ESMiddleware, index.GetSettings)
	r.PUT("/:target/_settings", ESMiddleware, index.SetSettings)

	r.POST("/_analyze", ESMiddleware, index.Analyze)
	r.POST("/:target/_analyze", ESMiddleware, index.Analyze)

	r.POST("/_aliases", ESMiddleware, index.AddOrRemoveESAlias)
	r.PUT("/_aliases", ESMiddleware, index.AddOrRemoveESAlias)

	r.GET("/_alias", ESMiddleware, index.GetESAliases)
	r.GET("/:target/_alias", ESMiddleware, index.GetESAliases)
	r.GET("/_alias/:target_alias", ESMiddleware, index.GetESAliases)

	// ES Bulk update/insert
	r.POST("/_bulk", ESMiddleware, document.ESBulk)
	r.PUT("/_bulk", ESMiddleware, document.ESBulk)
	r.POST("/:target/_bulk", ESMiddleware, document.ESBulk)
	r.PUT("/:target/_bulk", ESMiddleware, document.ESBulk)
	r.POST("/_refresh", index.RefreshAll)
	r.POST("/:target/_refresh", index.Refresh)
	// ES Document
	r.POST("/:target/_doc", ESMiddleware, document.CreateUpdate)        // create
	r.PUT("/:target/_doc/:id", ESMiddleware, document.CreateUpdate)     // create or update
	r.PUT("/:target/_create/:id", ESMiddleware, document.CreateUpdate)  // create
	r.POST("/:target/_create/:id", ESMiddleware, document.CreateUpdate) // create
	r.POST("/:target/_update/:id", ESMiddleware, document.Update)       // update part of document
	r.DELETE("/:target/_doc/:id", ESMiddleware, document.Delete)        // delete
	r.HEAD("/:target/_doc/:id", ESMiddleware, document.Get)             // get
	r.GET("/:target/_doc/:id", ESMiddleware, document.Get)              // get
}